# app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Medical Backend API",
        "description": "API for managing medical backend (Auth, Admin, Doctors)",
        "version": "1.0.0"
    },
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "paths": {
        "/auth/register": {
            "post": {
                "summary": "Register a new user",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "role": {
                                    "type": "string",
                                    "enum": [
                                        "admin",
                                        "doctor",
                                        "member"
                                    ]
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created"
                    },
                    "409": {
                        "description": "User already exists"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "summary": "Login user",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "token": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/admin/departments": {
            "get": {
                "summary": "List all departments",
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of departments"
                    }
                }
            },
            "post": {
                "summary": "Create a new department",
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Department created"
                    },
                    "409": {
                        "description": "Department already exists"
                    },
                    "403": {
                        "description": "Permission denied"
                    }
                }
            }
        },
        "/admin/doctors": {
            "post": {
                "summary": "Onboard a new doctor",
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Doctor onboarded"
                    },
                    "409": {
                        "description": "Doctor already exists"
                    }
                }
            }
        },
        "/admin/assign-doctor": {
            "post": {
                "summary": "Assign a doctor to a department",
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "doctor_id": {
                                    "type": "integer"
                                },
                                "department_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assignment successful"
                    },
                    "400": {
                        "description": "Invalid ID"
                    }
                }
            }
        }
    }
}